Fix type conversion error for before_xid_wraparound metric
Fixed error: "converting driver.Value type float64 to a int64: invalid syntax"

The PostgreSQL calculation (2^31 - age(datfrozenxid)) returns a numeric/float
type by default. Added explicit ::bigint cast to ensure the result is returned
as bigint, which correctly maps to sql.NullInt64 in Go.

Error was:
sql: Scan error on column index 16, name "before_xid_wraparound":
converting driver.Value type float64 ("2.14748336e+09") to a int64: invalid syntax

+++ receiver/newrelicpostgressqlreceiver/queries/database_metrics_queries.go
@@ -29,7 +29,7 @@ const (
 			sd.temp_bytes,
 			sd.blk_read_time,
 			sd.blk_write_time,
-			(2^31 - age(d.datfrozenxid)) AS before_xid_wraparound,
+			(2^31 - age(d.datfrozenxid))::bigint AS before_xid_wraparound,
 			pg_database_size(sd.datname) AS database_size
 		FROM pg_stat_database sd
 		JOIN pg_database d ON sd.datname = d.datname
